% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{complexity_measures}
\alias{complexity_measures}
\title{Complexity Measures}
\usage{
complexity_measures(bi = NULL, x = NULL, y = NULL, v = "value",
  method = "fitness", iterations = 20, extremality = 1, tbl = TRUE)
}
\arguments{
\item{bi}{A matrix or tibble/data.frame. You can pass the
output of \code{balassa_index()} here or a similar data structure.}

\item{x}{column with the elements of set X (applies only if d is a
tibble/data.frame)}

\item{y}{column with the elements of set Y (applies only if d is a
tibble/data.frame)}

\item{v}{column with a metric of the relation between the elements of the
sets X and Y (applies only if d is a tibble/data.frame)}

\item{method}{string to indicate to use one of these methods: reflections,
eigenvalues or fitness (default set to "fitness")}

\item{iterations}{number of iterations (default set to 20, all the methods
use this but "eigenvalues" uses it indirectly in sign correction)}

\item{extremality}{parameter to use in the fitness method
(default set to 1, the other methods ignore this parameter)}

\item{tbl}{when set to TRUE, the output will be a tibble instead of a
matrix (default set to TRUE)}
}
\description{
Complexity Measures
}
\examples{
complexity_measures(
  bi = binet_output$balassa_index, x = "country", y = "product", v = "value"
)

}
\references{
For more information on this index see:

\insertRef{measuringcomplexity2015}{binet}

and the references therein.
}
\keyword{functions}
