\name{galactic_federation}
\alias{galactic_federation}
\docType{data}
\title{
Fictional data based on Rick and Morty
}
\description{
Example dataset with fictional trade for the planets that were part of
the Galatic Federation in 2017 according to Rick and Morty. The planets are
taken from the TV show but the products and values are just an invention. Each
row expresses how many Blemflarcks (Galatic Federation currency) each planet
exported to the rest of the interplanetary trade system.
}
\usage{galactic_federation}
\format{
  A data frame with 108 observations on the following 3 variables.
  \describe{
    \item{\code{planet}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{export_value}}{a numeric vector}
  }
}
\source{
Open galactic_federation Statistics
}
\examples{
galactic_federation
}
\keyword{datasets}
