% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Projections of a bipartite network}
\usage{
projections(px = NULL, py = NULL, f = "from", t = "to",
  v = "value", ax = 4, ay = 4, tol = 0.01, tbl = TRUE,
  pro = "both")
}
\arguments{
\item{px}{tibble/data.frame or sparse/dense matrix containing proximity
measures for the elements of set X (e.g. proximity_x from \code{proximity()})}

\item{py}{tibble/data.frame or sparse/dense matrix containing proximity
measures for the elements of set Y (e.g. proximity_y from \code{proximity()})}

\item{f}{column with the "from" side
in a relation between elements of the same set (applies only if d is a
tibble/data.frame)}

\item{t}{column with the "to" side in a relation between elements of the same
set (applies only if d is a tibble/data.frame)}

\item{v}{column with a metric of the relation between the elements of the
same set (applies only if d is a tibble/data.frame)}

\item{ax}{average number of connections for the projection of X
(default set to 4)}

\item{ay}{average number of connections for the projection of Y
(default set to 4)}

\item{tol}{tolerance for proximity variation on each iteration until
obtaining the desired average number of connections (default set to 0.01)}

\item{tbl}{when set to TRUE the output will be a tibble instead of a
graph (set to TRUE by default)}

\item{pro}{set to compute "both" projections by default, it can also be
"x" or "y" to obtain one projection of the bipartite network}
}
\description{
Projections of a bipartite network
}
\examples{
projections(
  px = binet_output$proximity$proximity_x,
  py = binet_output$proximity$proximity_y,
  f = "from",
  t = "to",
  v = "value",
  tol = 0.25,
  tbl = TRUE
)

}
\references{
For more information see:

\insertRef{atlas2014}{binet}

and the references therein.
}
\keyword{functions}
