\name{trade}
\alias{trade}
\docType{data}
\title{
World trade in 1962
}
\description{
Example dataset with countries and the products they export. The exports are
grouped to four digits HS 2007 codes.
}
\usage{trade}
\format{
  A data frame with 44435 observations on the following 3 variables.
  \describe{
    \item{\code{country}}{a character vector}
    \item{\code{product}}{a character vector}
    \item{\code{export_value}}{a numeric vector}
  }
}
\source{
Open Trade Statistics
}
\examples{
trade
}
\keyword{datasets}
