% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balassa_index.R
\name{balassa_index}
\alias{balassa_index}
\title{Balassa Index (or Revealed Comparative Advantage, RCA)}
\usage{
balassa_index(d = NULL, x = NULL, y = NULL, v = NULL, cut = 1,
  dis = TRUE, tbl = TRUE)
}
\arguments{
\item{d}{tibble/data.frame in long format or a dense/sparse matrix}

\item{x}{column with the elements of set X (applies only if d is a
tibble/data.frame)}

\item{y}{column with the elements of set Y (applies only if d is a
tibble/data.frame)}

\item{v}{column with a metric of the relation between the elements of th
sets X and Y (applies only if d is a tibble/data.frame)}

\item{cut}{cutoff value used for discretization (default set
to 1)}

\item{dis}{when set to TRUE, all the Balassa Index values are converted
to discrete values, anything below the cutoff is converted to 0 and 1
otherwise (default set to TRUE)}

\item{tbl}{when set to TRUE, the output will be a tibble instead of a
matrix (default set to TRUE)}
}
\description{
Balassa Index (or Revealed Comparative Advantage, RCA)
}
\examples{
balassa_index(
  d = trade_1962, x = "country", y = "product", v = "value"
)

}
\references{
For more information see:

\insertRef{measuringcomplexity2015}{binet}

and the references therein.
}
\keyword{functions}
