% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balassa_index.R
\name{balassa_index}
\alias{balassa_index}
\title{Balassa Index}
\usage{
balassa_index(data, source = "source", target = "target",
  value = "value", discrete = TRUE, cutoff = 1)
}
\arguments{
\item{data}{a data frame}

\item{source}{a column with the elements of set X (applies only if data is
a data frame).}

\item{target}{a column with the elements of set Y (applies only if data is
a data frame).}

\item{value}{a column with some metric of the relation between the elements
of X and Y (applies only if data is a data frame).}

\item{discrete}{if \code{TRUE} the Balassa Index values are converted to
discrete values (1/0) unless you set this to \code{FALSE}. Anything below the
specified cutoff is converted to 0 and 1 otherwise.}

\item{cutoff}{the cutoff value used for discretization. The default is 1 but
it can be any numeric value.}
}
\description{
\code{balassa_index()} computes the Balassa Index for a
bipartite relation between two disjoint sets X (the "source" side) and Y (the
"target" side).
}
\details{
TBD
}
\examples{
balassa_index(
    data = trade,
    source = "country",
    target = "product",
    value = "export_value"
)

}
\references{
For more information see:

\insertRef{measuringcomplexity2015}{binet}

and the references therein.
}
\keyword{functions}
