% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{complexity_measures}
\alias{complexity_measures}
\title{Complexity Measures}
\usage{
complexity_measures(balassa_index, source = "source",
  target = "target", value = "value", method = "fitness",
  iterations = 20, extremality = 1)
}
\arguments{
\item{balassa_index}{a data frame or matrix (e.g. the output from
\code{balassa_index()}).}

\item{source}{a column with the elements of set X (applies only if data is
a data frame).}

\item{target}{a column with the elements of set Y (applies only if data is
a data frame).}

\item{value}{a column with some metric of the relation between the elements
of X and Y (applies only if data is a data frame).}

\item{method}{one of these methods: "fitness" (by default), "reflections",
"eigenvalues".}

\item{iterations}{the number of iterations to use (default set to 20).}

\item{extremality}{the parameter to use in the fitness method (by default is
1). The other methods ignore this parameter.}
}
\description{
\code{complexity_measures()} computes different complexity
measures obtained from the Balassa Index for a bipartite relation between
two disjoint sets X (the "source" side) and Y (the "target" side).
}
\details{
TBD
}
\examples{
complexity_measures(balassa_index = binet_output$balassa_index)

}
\references{
For more information on this index see:

\insertRef{measuringcomplexity2015}{binet}

and the references therein.
}
\keyword{functions}
