% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projections}
\alias{projections}
\title{Projections of a bipartite network}
\usage{
projections(proximity_source, proximity_target, source = "source",
  target = "target", value = "value", avg_links = 4,
  tolerance = 0.01, compute = "both")
}
\arguments{
\item{proximity_source}{tibble/data.frame or sparse/dense matrix containing proximity
measures for the elements of set X (e.g. proximity_x from \code{proximity()})}

\item{proximity_target}{tibble/data.frame or sparse/dense matrix containing proximity
measures for the elements of set Y (e.g. proximity_y from \code{proximity()})}

\item{source}{a column with the elements of set X (applies only if data is
a data frame).}

\item{target}{a column with the elements of set Y (applies only if data is
a data frame).}

\item{value}{a column with some metric of the relation between the elements
of X and Y (applies only if data is a data frame).}

\item{avg_links}{average number of connections for the projection of X
(default set to 4)}

\item{tolerance}{tolerance for proximity variation on each iteration until
obtaining the desired average number of connections (default set to 0.01)}

\item{compute}{which projection to compute. By default is "both" (both
projections) but it can also be "source" or "target".}

\item{tbl}{when set to TRUE the output will be a tibble instead of a
graph (set to TRUE by default)}
}
\description{
Projections of a bipartite network
}
\examples{
projections(
  proximity_source = binet_output$proximity$proximity_x,
  proximity_target = binet_output$proximity$proximity_y
)

}
\references{
For more information see:

\insertRef{atlas2014}{binet}

and the references therein.
}
\keyword{functions}
