% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Proximity}
\usage{
proximity(bi = NULL, x = "x", y = "y", v = "v", d = NULL,
  dx = NULL, dv = "value", u = NULL, uy = NULL, uv = "value",
  tbl = TRUE, mat = "both")
}
\arguments{
\item{bi}{A matrix or tibble/data.frame. You can pass the
output of \code{balassa_index()} here or a similar data structure.}

\item{x}{string to indicate the column that contains the elements of
set X (applies only if d is a tibble/data.frame)}

\item{y}{string to indicate the column that contains the elements of
set Y (applies only if d is a tibble/data.frame)}

\item{v}{string to indicate the column that contains a metric of the
relation between the elements of the sets X and Y (applies only
if d is a tibble/data.frame)}

\item{d}{numeric vector or tibble/data.frame containing diversity
measures (e.g. diversity from \code{complexity_measures()})}

\item{dx}{string to indicate the column that contains
the elements of set X in d (default set to NULL)}

\item{dv}{string to indicate the column that contains
a diversity metric for the elements of set X in d (default set to "value")}

\item{u}{numeric vector or tibble/data.frame containing ubiquity
measures (e.g. ubiquity from \code{complexity_measures()})}

\item{uy}{string to indicate the column that contains
the elements of set Y in u (default set to NULL)}

\item{uv}{string to indicate the column that contains
a ubiquity metric for the elements of set Y in u (default set to "value")}

\item{tbl}{when set to TRUE, the output will be a tibble instead of a
matrix (default set to TRUE)}

\item{mat}{set to compute "both" matrices by default, it can also be
"x" or "y" to obtain only one matrix from the bipartite relation}
}
\description{
Proximity
}
\examples{
proximity(
  bi = binet_output$balassa_index,
  x = "country",
  y = "product",
  v = "value",
  d = binet_output$complexity_measures$diversity,
  dx = "country",
  dv = "value",
  u = binet_output$complexity_measures$ubiquity,
  uy = "product",
  uv = "value"
)
}
\references{
For more information see:

\insertRef{atlas2014}{binet}

and the references therein.
}
\keyword{functions}
